requirements.txt
numpy>=1.24
pandas>=2.0
matplotlib>=3.7
pyyaml>=6.0
streamlit>=1.30
scipy>=1.11
pytest>=7.4
supabase>=2.0
python-dotenv>=1.0